// BuyerShah/project/src/features/services/DTOs/interfaces.ts
export interface ProductFilterRequest {
  categoryId?: string | null;
  includeChildCategories?: boolean;
  page?: number;
  pageSize?: number;
  minPrice?: number | null;
  maxPrice?: number | null;
  valueIds?: string[];
  userId?: string | null;
}